+++ src/components/Group/Name.vue
@@ -2,7 +2,7 @@
   <FormButton
     data-bs-toggle="modal"
     data-bs-target="#group-edit-modal"
-    class="btn-link truncate text-left"
+    class="truncate text-left"
   >
     <FontAwesomeIcon icon="layer-group" class="mr-4" /><slot></slot>
   </FormButton>