+++ src/components/Note/Search.vue
@@ -1,26 +1,22 @@
 <template>
   <form class="note-search" novalidate @submit="onSubmit">
-    <div class="input-group mb-4">
+    <div class="input-group px-4 pt-4">
       <input
         v-model="searchInput"
         aria-label="Search Notes"
         aria-placeholder="Search Notes"
-        class="search-input form-control rounded-full"
+        class="search-input form-control rounded-md p-2"
         placeholder="Search"
         type="text"
         @blur="onBlur"
+        @input="onChange"
       />
-      <button
-        v-if="searchInput"
-        class="btn btn-primary rounded-full"
-        type="button"
-        @click="onClear"
-      >
-        <FontAwesomeIcon icon="times" />
-      </button>
-      <button class="btn btn-primary rounded-full px-4" type="submit">
-        <FontAwesomeIcon icon="search" />
-      </button>
+      <FormButton v-if="searchInput" @click="onClear">
+        <FontAwesomeIcon class="mx-2" icon="times" />
+      </FormButton>
+      <FormButton type="submit">
+        <FontAwesomeIcon class="mx-2" icon="search" />
+      </FormButton>
     </div>
   </form>
 </template>
@@ -31,6 +27,13 @@ import { ref } from 'vue';
 const emit = defineEmits(['onSearch']);
 
 const searchInput = ref('');
+
+function onChange(event) {
+  event.preventDefault();
+  if (event.target.value === '') {
+    emit('onSearch', '');
+  }
+}
 
 function onSubmit(event) {
   event.preventDefault();