+++ src/components/Tags.vue
@@ -2,12 +2,9 @@
   <div
     class="
       flex justify-between items-center
-      border-t-2 border-indigo-500
     "
   >
-    <h1 class="text-lg py-4 px-8">
-      <FontAwesomeIcon icon="tags" class="mr-4" />Tags
-    </h1>
+    <h1><FontAwesomeIcon icon="tags" class="mr-4" />Tags</h1>
   </div>
 
   <div class="overflow-y-auto flex-1">
@@ -15,14 +12,11 @@
 
     <div v-else-if="error">{{ error }}</div>
 
-    <div v-else class="text-sm">
+    <div v-else>
       <div
         v-for="tag in tags?.listTags.items"
         :key="tag.id"
         class="
-        py-3
-        px-6
-        mx-2
         cursor-pointer
         hover:bg-pink-800
         flex