+++ src/pages/index.vue
@@ -3,26 +3,21 @@
     class="
       h-full
       text-white
-      bg-gradient-to-r
+      bg-gradient-to-t
       from-pink-600
       via-purple-700
-      to-blue-700
+      to-indigo-700
     "
   >
-    <div
-      class="
-        grid grid-cols-1
-        lg:grid-cols-3
-        xl:grid-cols-4
-        2xl:grid-cols-5
-        h-full
-        bg-gradient-to-b
-        from-gray-900
-        via-purple-900
-        to-transparent
-      "
-    >
+    <div class="flex h-full">
       <Groups />
+
+      <Group :key="$route.params.groupId" v-if="$route.params.groupId" />
+
+      <Note
+        v-if="$route.params.noteId"
+        :key="`${$route.params.groupId}-${$route.params.noteId}`"
+      />
     </div>
   </div>
 </template>
@@ -44,12 +39,10 @@ import {
   didAuthError,
 } from '../urql/authExchange';
 
-console.log(process.env);
-
 const client = createClient({
   url:
     process.env.NODE_ENV !== 'production'
-      ? process.env.VUE_APP_API_ENDPOINT
+      ? process.env.AMPLIFY_MOCK_ENDPOINT
       : 'https://qfzdtvzgdje3tbvnk63mwi5dta.appsync-api.us-east-1.amazonaws.com/graphql',
   exchanges: [
     authExchange({