+++ tailwind.config.js
@@ -1,10 +1,14 @@
-// tailwind.config.js
 const svgToDataUri = require('mini-svg-data-uri');
 
 module.exports = {
   purge: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
-  darkMode: false, // or 'media' or 'class'
   theme: {
+    screens: {
+      sm: '480px',
+      md: '768px',
+      lg: '976px',
+      xl: '1440px',
+    },
     minWidth: {
       '0': '0',
       '1/4': '25%',